Move of code into a folder
